+++ src/arcane_arcade_spelleffects/WaveEffect.java
@@ -50,7 +50,7 @@ public class WaveEffect extends MaskedSpellEffect
 	//IMPLEMENTED METHODS----------------------------------------------
 	
 	@Override
-	public void onBallCollision(Ball ball, int x, int y)
+	public void onBallCollision(Ball ball, double x, double y)
 	{
 		// Causes impact to the ball
 		ball.impact(2.2 * getForceModifier(ball), 4, getMovement().getDirection(), 
@@ -58,12 +58,12 @@ public class WaveEffect extends MaskedSpellEffect
 	}
 
 	@Override
-	public void onSpellCollision(SpellEffect spell, int x, int y) {
+	public void onSpellCollision(SpellEffect spell, double x, double y) {
 		// Doesn't collide with spells
 	}
 
 	@Override
-	public void onWizardCollision(Wizard wizard, int x, int y) {
+	public void onWizardCollision(Wizard wizard, double x, double y) {
 		// Doesn't collide with Wizards
 	}
 }
